Split the retrieval of domain commands into separate classes. DefaultCustomDomainCommandResolver now inherits from AggregatedCustomDomainCommandResolver and returns known resolvers by default. This allows new domain command resolvers to be easily added without the resolver becoming bulky.
