+++ MFiles.VAF.Extensions.Tests/Dashboards/Commands/CustomDomainCommandResolution/AttributeCustomDomainCommandResolverTests.cs
@@ -1,6 +1,7 @@
 ﻿using MFiles.VAF.Configuration.AdminConfigurations;
 using MFiles.VAF.Configuration.Domain;
 using MFiles.VAF.Extensions.Dashboards.Commands;
+using MFiles.VAF.Extensions.Dashboards.Commands.CustomDomainCommandResolution;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Collections.Generic;
@@ -8,10 +9,10 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MFiles.VAF.Extensions.Tests.Dashboards.Commands
+namespace MFiles.VAF.Extensions.Tests.Dashboards.Commands.CustomDomainCommandResolution
 {
 	[TestClass]
-	public class DefaultCustomDomainCommandResolverTests
+	public class AttributeCustomDomainCommandResolverTests
 	{
 		/// <summary>
 		/// By default there should be no commands returned.
@@ -19,7 +20,7 @@ namespace MFiles.VAF.Extensions.Tests.Dashboards.Commands
 		[TestMethod]
 		public void NoIncludedTypesReturnsEmptyCollection()
 		{
-			var resolver = new DefaultCustomDomainCommandResolver();
+			var resolver = new AttributeCustomDomainCommandResolver();
 			var commands = resolver.GetCustomDomainCommands();
 			Assert.IsNotNull(commands);
 			Assert.AreEqual(false, commands.Any());
@@ -38,7 +39,7 @@ namespace MFiles.VAF.Extensions.Tests.Dashboards.Commands
 		[TestMethod]
 		public void InvalidMethodDoesNotThrow()
 		{
-			var resolver = new DefaultCustomDomainCommandResolver();
+			var resolver = new AttributeCustomDomainCommandResolver();
 			resolver.Include(new InvalidInstanceMethod());
 			var commands = resolver.GetCustomDomainCommands();
 			Assert.IsNotNull(commands);
@@ -57,7 +58,7 @@ namespace MFiles.VAF.Extensions.Tests.Dashboards.Commands
 		[TestMethod]
 		public void ValidInstanceMethod_ReturnsValidData()
 		{
-			var resolver = new DefaultCustomDomainCommandResolver();
+			var resolver = new AttributeCustomDomainCommandResolver();
 			resolver.Include(new ValidInstanceMethod());
 			var commands = resolver.GetCustomDomainCommands();
 			Assert.IsNotNull(commands);
@@ -66,7 +67,7 @@ namespace MFiles.VAF.Extensions.Tests.Dashboards.Commands
 			var command = commands.ElementAt(0);
 			Assert.IsNotNull(command);
 			Assert.AreEqual("hello world", command.DisplayName);
-			Assert.AreEqual("MFiles.VAF.Extensions.Tests.Dashboards.Commands.DefaultCustomDomainCommandResolverTests+ValidInstanceMethod.Method", command.ID);
+			Assert.AreEqual("MFiles.VAF.Extensions.Tests.Dashboards.Commands.CustomDomainCommandResolution.AttributeCustomDomainCommandResolverTests+ValidInstanceMethod.Method", command.ID);
 			Assert.AreEqual(default, command.Blocking);
 			Assert.AreEqual(default, command.ConfirmMessage);
 			Assert.AreEqual(default, command.HelpText);
@@ -90,7 +91,7 @@ namespace MFiles.VAF.Extensions.Tests.Dashboards.Commands
 		[TestMethod]
 		public void ValidInstanceMethodAndAdditionalData_ReturnsValidData()
 		{
-			var resolver = new DefaultCustomDomainCommandResolver();
+			var resolver = new AttributeCustomDomainCommandResolver();
 			resolver.Include(new ValidInstanceMethodAndAdditionalData());
 			var commands = resolver.GetCustomDomainCommands();
 			Assert.IsNotNull(commands);
@@ -99,7 +100,7 @@ namespace MFiles.VAF.Extensions.Tests.Dashboards.Commands
 			var command = commands.ElementAt(0);
 			Assert.IsNotNull(command);
 			Assert.AreEqual("hello world", command.DisplayName);
-			Assert.AreEqual("MFiles.VAF.Extensions.Tests.Dashboards.Commands.DefaultCustomDomainCommandResolverTests+ValidInstanceMethodAndAdditionalData.Method", command.ID);
+			Assert.AreEqual("MFiles.VAF.Extensions.Tests.Dashboards.Commands.CustomDomainCommandResolution.AttributeCustomDomainCommandResolverTests+ValidInstanceMethodAndAdditionalData.Method", command.ID);
 			Assert.AreEqual(true, command.Blocking);
 			Assert.AreEqual("abc", command.ConfirmMessage);
 			Assert.AreEqual("def", command.HelpText);
@@ -123,7 +124,7 @@ namespace MFiles.VAF.Extensions.Tests.Dashboards.Commands
 		[TestMethod]
 		public void ValidInstanceMethodWithCustomCommandId_ReturnsValidData()
 		{
-			var resolver = new DefaultCustomDomainCommandResolver();
+			var resolver = new AttributeCustomDomainCommandResolver();
 			resolver.Include(new ValidInstanceMethodWithCustomCommandId());
 			var commands = resolver.GetCustomDomainCommands();
 			Assert.IsNotNull(commands);
@@ -147,7 +148,7 @@ namespace MFiles.VAF.Extensions.Tests.Dashboards.Commands
 		[TestMethod]
 		public void ValidStaticMethod_ReturnsValidData()
 		{
-			var resolver = new DefaultCustomDomainCommandResolver();
+			var resolver = new AttributeCustomDomainCommandResolver();
 			resolver.Include(new ValidStaticMethod());
 			var commands = resolver.GetCustomDomainCommands();
 			Assert.IsNotNull(commands);
@@ -156,7 +157,7 @@ namespace MFiles.VAF.Extensions.Tests.Dashboards.Commands
 			var command = commands.ElementAt(0);
 			Assert.IsNotNull(command);
 			Assert.AreEqual("hello world", command.DisplayName);
-			Assert.AreEqual("MFiles.VAF.Extensions.Tests.Dashboards.Commands.DefaultCustomDomainCommandResolverTests+ValidStaticMethod.Method", command.ID);
+			Assert.AreEqual("MFiles.VAF.Extensions.Tests.Dashboards.Commands.CustomDomainCommandResolution.AttributeCustomDomainCommandResolverTests+ValidStaticMethod.Method", command.ID);
 			Assert.AreEqual(default, command.Blocking);
 			Assert.AreEqual(default, command.ConfirmMessage);
 			Assert.AreEqual(default, command.HelpText);
@@ -179,7 +180,7 @@ namespace MFiles.VAF.Extensions.Tests.Dashboards.Commands
 		[TestMethod]
 		public void ValidInstanceMethod_Overridden_ReturnsValidData()
 		{
-			var resolver = new DefaultCustomDomainCommandResolver();
+			var resolver = new AttributeCustomDomainCommandResolver();
 			resolver.Include(new ValidInstanceMethod_Overridden());
 			var commands = resolver.GetCustomDomainCommands();
 			Assert.IsNotNull(commands);
@@ -192,7 +193,7 @@ namespace MFiles.VAF.Extensions.Tests.Dashboards.Commands
 			// NOTE: the command ID is changed by the overriding.
 			// This is "expected" (as the method is now the overriding one), but is it actually expected?
 			// For this reason people should not make assumptions about the command ID, but retrieve it from...  Something?
-			Assert.AreEqual("MFiles.VAF.Extensions.Tests.Dashboards.Commands.DefaultCustomDomainCommandResolverTests+ValidInstanceMethod_Overridden.Method", command.ID);
+			Assert.AreEqual("MFiles.VAF.Extensions.Tests.Dashboards.Commands.CustomDomainCommandResolution.AttributeCustomDomainCommandResolverTests+ValidInstanceMethod_Overridden.Method", command.ID);
 
 			Assert.AreEqual(default, command.Blocking);
 			Assert.AreEqual(default, command.ConfirmMessage);
@@ -217,7 +218,7 @@ namespace MFiles.VAF.Extensions.Tests.Dashboards.Commands
 		[TestMethod]
 		public void ValidInstanceMethod_OverriddenWithNewAttribute_ReturnsValidData()
 		{
-			var resolver = new DefaultCustomDomainCommandResolver();
+			var resolver = new AttributeCustomDomainCommandResolver();
 			resolver.Include(new ValidInstanceMethod_OverriddenWithNewAttribute());
 			var commands = resolver.GetCustomDomainCommands();
 			Assert.IsNotNull(commands);
@@ -226,7 +227,7 @@ namespace MFiles.VAF.Extensions.Tests.Dashboards.Commands
 			var command = commands.ElementAt(0);
 			Assert.IsNotNull(command);
 			Assert.AreEqual("overridden label", command.DisplayName);
-			Assert.AreEqual("MFiles.VAF.Extensions.Tests.Dashboards.Commands.DefaultCustomDomainCommandResolverTests+ValidInstanceMethod_OverriddenWithNewAttribute.Method", command.ID);
+			Assert.AreEqual("MFiles.VAF.Extensions.Tests.Dashboards.Commands.CustomDomainCommandResolution.AttributeCustomDomainCommandResolverTests+ValidInstanceMethod_OverriddenWithNewAttribute.Method", command.ID);
 
 		}
 	}
